feat:增加mpc_pro,增加dlyap函数

+++ include/LQR.h
@@ -27,20 +27,20 @@ public:
     int max_num_iteration;
     float TS;
     int dim_x, dim_u;
-    Eigen::MatrixXf Q;
-    Eigen::MatrixXf R;
-    Eigen::MatrixXf P;
-    Eigen::MatrixXf P_next;
-    Eigen::MatrixXf K;
-    Eigen::MatrixXf A_d;
-    Eigen::MatrixXf B_d;
+    Eigen::MatrixXd Q;
+    Eigen::MatrixXd R;
+    Eigen::MatrixXd P;
+    Eigen::MatrixXd P_next;
+    Eigen::MatrixXd K;
+    Eigen::MatrixXd A_d;
+    Eigen::MatrixXd B_d;
 
     LQR(std::string _name);
     ~LQR();
-    void get_param(Eigen::MatrixXf _Q, Eigen::MatrixXf _R, float ts);
-    void discrete(Eigen::MatrixXf &_A, Eigen::MatrixXf &_B, int type);
-    void compute_ARE(Eigen::MatrixXf &A, Eigen::MatrixXf &B, bool is_dis);
-    bool is_controllable(Eigen::MatrixXf &A, Eigen::MatrixXf &B);
+    void get_param(Eigen::MatrixXd _Q, Eigen::MatrixXd _R, float ts);
+    void discrete(Eigen::MatrixXd &_A, Eigen::MatrixXd &_B, int type);
+    void compute_ARE(Eigen::MatrixXd &A, Eigen::MatrixXd &B, bool is_dis);
+    bool is_controllable(Eigen::MatrixXd &A, Eigen::MatrixXd &B);
 };
 
 #endif
